backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic-settings==2.4.0
httpx==0.27.2
PyYAML==6.0.2
python-dotenv==1.0.1
